requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.41.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
cmake==3.29.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
dask==2023.3.1
dataclass_array==1.5.1
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.9.5
descartes==1.1.0
diskcache==5.6.3
dm-tree==0.1.8
dnspython==2.6.1
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
einsum==0.3.0
email_validator==2.2.0
entrypoints==0.4
etils==1.10.0
evdev==1.7.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gast==0.4.0
gguf==0.6.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.35.2
gradio_client==0.8.1
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.36.0
immutabledict==2.2.0
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.22.2
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.7.2
keras==2.12.0
keyboard==0.13.5
kiwisolver==1.4.5
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
lit==18.1.8
llama_cpp_python==0.2.79
llamacpp==0.1.14
LLM-SRP==0.1
locket==1.0.0
Markdown==3.7
markdown-it-py==2.2.0
markdown2==2.4.8
MarkupSafe==2.1.5
matplotlib==3.6.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==1.23.0
nuscenes-devkit==1.1.11
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
open_clip_torch==2.26.1
open-flamingo==2.0.1
opencv-python==4.10.0.84
OpenEXR==1.3.9
opt_einsum==3.4.0
orjson==3.10.5
packaging==24.0
pandas==1.5.3
parso==0.8.4
partd==1.4.2
peft==0.4.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.2.0
pip==23.3.1
platformdirs==4.2.1
plotly==5.13.1
promise==2.3
prompt-toolkit==3.0.42
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==10.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pydantic==1.10.13
pydantic_core==2.20.0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
pyquaternion==0.9.9
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-xlib==0.33
pytz==2024.1
PyWavelets==1.7.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
regex==2024.5.15
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.5.0
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.98
setuptools==67.6.0
Shapely==1.8.5.post1
shellingham==1.5.4
shortuuid==1.0.11
simple-parsing==0.1.6
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.6
tensorflow-estimator==2.12.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-probability==0.19.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.9.12
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
trimesh==4.5.0
triton==2.0.0
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.22.0
uvloop==0.19.0
visu3d==1.5.1
watchfiles==0.22.0
wavedrom==2.0.3.post3
waymo-open-dataset-tf-2-12-0==1.6.4
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.4
wheel==0.41.2
wrapt==1.14.1
yarl==1.9.4
zipp==3.20.2
